% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgClust.R
\name{sgClust}
\alias{sgClust}
\title{sparse Gaussian mixture model .}
\usage{
sgClust(
  data,
  lambda,
  c_center = NULL,
  v_int = NULL,
  mu = 0,
  Sigma = 1,
  pi_int = NULL,
  K = 2,
  no_init = 1,
  max_iter = 200
)
}
\arguments{
\item{data}{data matrix, where rows represent genes and columns represents samples.}

\item{lambda}{penalty parameter needs to be tuned or specified.}

\item{c_center}{Initial centers for each clusters. We suggest using
other clustering algorithms such as sparse K-means to speed up.}

\item{K}{number of clusters.}
}
\value{
A list of two components:
\itemize{
\item{result: }{a list containing mean matrix (mu),variable estimation for each feature (sigma)
likelihood, cluster assignment probability (z) and prior probability for each cluster (pi)}
\item{BIC: }{ Extended BIC of the result}

}
}
\description{
sparse Gaussian mixture model
}
\references{
Pan W, Shen X. Penalized model-based clustering with application to variable selection[J]. Journal of Machine Learning Research, 2007, 8(May): 1145-1164.
}
