% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Corr.R
\name{Sim.Corr}
\alias{Sim.Corr}
\title{Simulation 1}
\usage{
Sim.Corr(alpha = 0.5, sim_disp, empirical_dist)
}
\arguments{
\item{alpha}{gene correlation parameter, larger alpha creates larger correlation}

\item{sim_disp}{gene dispersion from breast cancer data}

\item{empirical_dist}{average gene expression from breast cancer data}
}
\value{
data with correlated genes for simulation 3.
}
\description{
Simulation 1
}
\references{
Rahman T, Li Y, Ma T, et al. A sparse negative binomial mixture model for clustering RNA-seq count data[J]. arXiv preprint arXiv:1912.02399, 2019.
}
