% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnbClust.R
\name{snbClust}
\alias{snbClust}
\title{sparse negative binomial mixture model for clustering RNA-seq count data.}
\usage{
snbClust(
  data,
  phi = rep(1, dim(data)[1]),
  lib,
  c_center = NULL,
  k = 2,
  penalize = FALSE,
  tune = 0,
  no_init = 1,
  max_iter_beta = 200,
  max_iter = 100,
  lambda = 100,
  m_node = F
)
}
\arguments{
\item{data}{data matrix, where rows represent genes and columns represents samples.}

\item{phi}{dispersion parameter for each gene.}

\item{lib}{library size normalization factor for each sample.}

\item{k}{number of clusters.}

\item{tune}{penalty parameter needs to be tuned or specified.}

\item{center}{Initial centers for each clusters. We suggest using
other clustering algorithms such as sparse K-means to speed up.}
}
\value{
A list of two components:
\itemize{
\item{result: }{a list containing mean matrix (beta),
likelihood, cluster assignment probability (z) and prior probability for each cluster (pi)}
\item{BIC: }{ Extended BIC of the result}

}
}
\description{
sparse negative binomial mixture model for clustering RNA-seq count data
}
\references{
Rahman T, Li Y, Ma T, et al. A sparse negative binomial mixture model for clustering RNA-seq count data[J]. arXiv preprint arXiv:1912.02399, 2019.
}
